Render thumbnails for point/pin annotations
Point annotations don't specify how much of the image the thumbnail should
cover. As a simple heuristic, use a square area that is 10% of the page width in
size, centered on the annotated point. The location of the point within the
thumbnail region is indicated by a dot drawn on the thumbnail.