Fix thumbnail rendering if aspect-ratio scaled height is zero
Fix thumbnail rendering of very wide rectangles such that, after scaling the
width to fit the maximum width, the height is less than one. In this case the
`OffscreenCanvas` constructor rounded the height down to zero and calling
`transferToImageBitmap` on a zero-height canvas would either fail (in Chrome) or
return `null` (in Safari).