Use percentage positioning for PDF text highlights
PDF text highlights are created using an `<svg>` positioned over the page's
`<canvas>`. Use percentages rather than pixel values to position the `<rect>`s
in this SVG layer. This keeps the highlights in the correct position when the
`<canvas>` and `<svg>` change size as a result of zooming.

This fixes an issue with highlights becoming misplaced at high zoom levels,
described in
https://github.com/hypothesis/client/issues/7015#issuecomment-2838091196.