Improve handling of links inside PDFs
It is annoying when accidentally clicking a link inside a PDF navigates away
from the PDF and unloads Hypothesis. This is more likely when text containing a
link is highlighted. In the LMS context this is especially problematic because
the user doesn't have a way to navigate back.

Remedy this by:

 - Making all links open in a new tab when the Hypothesis PDF
   integration is active.
 - Disabling default behavior for highlighted links, when the highlights
   are visible. Clicking on a highlighted link will focus the highlight
   rather than following the link. When highlights are hidden, the link
   will become active but open in a new tab.

Fixes https://github.com/hypothesis/support/issues/181