Make pin drawing finish on mouse up rather than mouse down
Make pin drawing work more like rect drawing, where selection finishes on
mouse up rather than mouse down. When the mouse is initially pressed and while
dragging it afterwards, show the pending location of the new pin as a filled
circle.

This allows the user to fine-tune the position of the pin after pressing the
mouse. Making rect and point annotations work in the same way also makes it a
bit easier to reason about event handling elsewhere in the code.