Fix annotating when even/odd page spreads are used
When creating an annotation the index of the page containing the text layer was
determined by getting the index of the containing `.page` element amongst its
siblings. This doesn't work when the Even Spreads or Odd Spreads layout modes
are used as the DOM structure changes and the `.page` elements are no longer all
under one parent. The solution is to get the page number from the
`data-page-number` attribute on the element instead.

Fixes https://github.com/hypothesis/support/issues/192