Take PDF page border into account when creating shape selectors
The border of the `.page` container elements was not taken into account when
generating shape selectors for rect/pin annotations. This resulted in the
coordinates of the shape being slightly down and to the right of where they
should have been. This was corrected for when drawing highlights by subtracting
the border size, but the error was actually when capturing the shape coordinates
in PDF space. As a result the area shown in thumbnails was slightly offset from
where the user clicked.