Make `compareByLocation` return a consistent ordering
Previously `compareByLocation` would compare annotations by different fields
depending on which selectors the annotations had. When sorting a list of
annotations with different selectors, this could lead to inconsistent ordering.

For example: Given PDF text annotations A and B, where A < B by text position,
and rect annotation C, if they are created in the order B C A,
`compareByLocation` would return the following orderings:

  A < B (ordered by char offset)
  B < C (ordered by creation date)
  C < A (ordered by creation date)

This is inconsistent, and so the final ordering produced by
`Array.prototype.sort` is unpredictable.

The fix is to always compare by the same fields, but use default values if a
field is missing. For example instead of skipping comparison by page number if
only one annotation has a page number, we instead fill in a page number of zero
for the other annotation.