Ensure adder is displayed above shape highlights
The existing logic for making the adder show above elements in the selected text
did not work because shape highlights have `pointer-events: none` set on them,
and `document.elementsFromPoint` ignores these.

There is no API to hit test elements with pointer events disabled, so we either
have to test all elements in the document, or as in this commit, query for
specific elements to test.

Fixes https://github.com/hypothesis/client/issues/7052