Make toolbar buttons taller when using touch input
Make the annotation toolbar buttons taller on mobile to reduce taps triggering
the wrong button. Ideally the buttons would be made wider as well, but this
requires some additional work to enable making the whole toolbar wider. As noted
in comments, the 40px height is slightly smaller than we normally use for touch
targets, but works in this context and also aligns with the toggle button that
opens and closes the sidebar.