Limit size of thumbnail cache
Add a size limit to the thumbnail cache and expire the least-recently-used
entries when new thumbnails are fetched. Thumbnails are moved to the back of the
cache's key list (ie. marked as most-recently-used) when the thumbnail is first
added, or an entry is fetched for use in an `AnnotationThumbnail` via calls to
`get` or `fetch`.