Replace "Search annotations..." with just "Search..."
For bioRxiv we're going to have to implement a "comments-only" mode
where all references to "annotation(s)" in the client's UI are hidden or
changed to refer to "comments" (see https://github.com/hypothesis/client/issues/7065).

This PR simply replaces the "Search annotations..." placeholder in the
search field (and various labels etc) with just "Search...". If
acceptable this should be simpler than having a config setting to change
it to "Search comments..."

I think the "annotations" / "comments" in these labels and placeholders
is probably unnecessary: "search" on its own should be clear enough?