Update browser baseline
The new baseline corresponds to the current requirements for the PDF.js legacy
build. See
https://github.com/hypothesis/client/issues/6784#issuecomment-2883162200.

The changes in the generated code are fairly small. The annotator bundle is
unchanged and the sidebar bundle is 1KB smaller (in production) due to removal
of some transpilation (eg. of private fields) that is no longer required.