Warn when closing tab if there is an unsaved, non-empty draft
Copy the `useUnsavedChanges` hook from the h repository and use it in
AnnotationEditor to enable a warning if the user tries to close the tab while
there is a non-empty draft. In the process, fix a bug where `returnValue` was
set to an empty string when according to MDN, it must be set to a truthy value.
Without this change, the hook works in Chrome but not Firefox.

This change works in Chrome and Firefox, but not Safari. In Safari, the tab
closes even when the `beforeunload` handler is active.

Part of https://github.com/hypothesis/support/issues/59.