Do not show group type icons for third-party authorities
In the LMS app, the group type icons were confusing for students since they
didn't join the groups themselves (the LMS app added them) and thus they don't
have a notion of having joined a private group, plus all the groups are private
anyway, so the icon isn't informative.

Group type icons are primarily intended for use with the first-party authority.
We might need to make it possible for third-party authorities to enable these
icons in future, but for the moment hiding the icons for all third-party
authorities is the simplest thing to do.

Fixes https://github.com/hypothesis/support/issues/183