Change how boot script gets asset root URLs
Previously the boot script generated the root URLs for assets using:

```
processUrlTemplate(config.assetRoot || 'https://cdn.hypothes.is/...')
```

Where the "cdn.hypothes.is" URL is used in the embed and `config.assetRoot` is
set in the extension, so our CDN URL is not used there. The presence of the CDN
URL in the extension caused Chrome Web Store reviewers to think the extension
was loading remote code, even though it isn't.

To resolve this, the boot script build is being changed to generate two separate
files, `build/boot-template.js` and `build/boot.js`. The template will contain
code like `processUrlTemplate(config.assetRoot || '__ASSET_ROOT_')` and
`build/boot.js` is a "rendered" version of the template with variables replaced
by CDN URLs. In the extension we will replace `boot.js` with a different
rendered version of the template, containing extension-specific URLs. This way
the CDN URL will not appear in the extension's scripts.

The `assetRoot` and related config keys are still supported because they are
used when Hypothesis injects itself into a same-origin iframe that has an
`enable-annotation` attribute. Except for that, we could remove them entirely.