Handle Unicode normalization differences in `translateOffsets`
In current versions of PDF.js, text extracted from the PDF via the PDF.js APIs
is normalized using a subset of NFKC normalization, as was the case for older
versions of PDF.js, whereas text in the text layer is not. This causes
misalignments if we try to use anchor positions in extracted text to create DOM
ranges that reference positions in the text layer.

We had a `translateOffsets` function which handled this problem for whitespace,
extend it to also handle Unicode normalization. The translation works by
normalizing both input strings before computing the alignment between a range in
the input string and a range in the output string. Once the alignment is
computed the positions are mapped back into positions in the un-normalized
string.

This implementation has no caching, which we might need to add if decomposition
proves expensive.

See https://github.com/hypothesis/client/issues/6784#issuecomment-2624697567.