Add "more info" icon to thumbnail description field
Extract the thumbnail description field into a new
`AnnotationThumbnailDescription` component and add an info icon button to the
right of the input. When clicked, a popover is displayed that explains briefly
what the description is used for, with a link to a KB article that explains in
more detail. The placeholder for the field has been changed to make the
distinction with the annotation comment field more obvious.