Use `<img>` rather than `<canvas>` to render thumbnails
The decision to use `<canvas>` elements rather than `<img>`s to render
thumbnails was motivated by performance (avoids overhead of serializing the
thumbnail as a PNG image) and avoiding the need to release thumbnail URLs when
no longer used.

However this has a downside of losing various affordances that browsers provide
for images such as being able to copy the image from the context menu or select
text in them (in Safari). On reflection it seems a better choice to use an
`<img>` and preserve these capabilities.