Don't override text color when styled_highlight_clusters is enabled
The `styled_highlight_clusters` feature is a prototype for supporting multiple
highlight colors on a page using a palette. When enabled it overrides the text
color in highlights to ensure contrast between the background and text. This
causes problems however for web pages which use color to distinguish links or
other clickable text, as the link color gets overridden to be the same as other
text in the highlight. Using color alone to distinguish links is a bad practice,
but many real web pages do this.

This commit resolves the issue by removing the `color` override. Instead we just
rely on our highlight colors working well against most text, as we do when the
feature is off. This generally works well for pages with light backgrounds. For
pages with dark backgrounds the color override is helpful, but text is still
readable, albeit with less clarity, without it.

An alternative would be to force underlining of links inside highlights. This
wouldn't help with other semantically important uses of color inside highlighted
text however.