Fix share action being shown incorrectly in LMS app
This is a partial revert of https://github.com/hypothesis/client/pull/7186.
Prior to that PR the annotation share control was hidden in two cases:

 1. Sharing is globally disabled for the client. This is true in the LMS.
 2. The annotation does not have a share URL

Case (2) is handled by the `AnnotationShareControl` component in the
@hypothesis/annotation-ui package, but the checks for case (1) were removed and
that still needs to be handled in the Hypothesis client.

Fixes https://github.com/hypothesis/support/issues/199