Send notification to LMS frontend when there are unsaved changes
We want to prevent accidentally losing unsaved changes by closing the tab. In
Chrome and Firefox we do this using `beforeunload` in the client. In desktop
Safari this does not work if the `beforeunload` handler is set in a cross-origin
iframe. To work around this, have the client send a message to the LMS frontend
which can install/remove a `beforeunload` handler in the top-level frame
instead. In mobile Safari this change doesn't help because `beforeunload` is
ignored there entirely.

See https://github.com/hypothesis/support/issues/59#issuecomment-3068704178