Increase the height of the login popup window
Increase the height of the login popup window to accomodate the UIs of the
different social login providers (Google, Facebook, ORCID) that the user may
directed through. This also creates room to show social login options on the
initial form.

For a short period of time the window will seem redundantly tall, until we
enable social login in production.

See
https://hypothes-is.slack.com/archives/C4K6M7P5E/p1754045015491359?thread_ts=1754040814.938669&cid=C4K6M7P5E
for notes.