Make "Sign up" link open a popup window
Change the "Sign up" link to open the same popup window as the "Log in" link,
but with an `action=signup` parameter. This allows the user to sign up via
social login and get logged into the client immediately.

See https://github.com/hypothesis/h/pull/9817 for the backend changes.