Fix adder Z-index not being set in some cases
Fix an issue where the `z-index` style property of the
`<hypothesis-adder>` element was not set if any of the reference
elements used to determine the value had a computed z-index of "auto".

If a reference element did have a computed z-index of "auto", then the
`elementZIndex` function would attempt to convert this to a number using
`+zIndex` which produced NaN. The NaN would be returned from `_findZIndex` and
was then ignored when assigned to `element.style.zIndex`. The fix is to treat
"auto" values as zero.

Related #support thread: https://hypothes-is.slack.com/archives/C2BLQDKHA/p1756497119701239