Add hack to work around issue with shadow DOM and @property
Add a hack to work around
https://github.com/tailwindlabs/tailwindcss/issues/15005. @property
declarations are ignored by current browsers in Shadow DOM. Tailwind v4's
styles rely on these properties. In the Hypothesis client this noticeably
affects borders of elements in the sidebar container and shadows on the adder.

This commit works around this by duplicating @property declarations from
annotator.css inside the main document. The declarations are hardcoded as a
string in shadow-root.ts on the assumption they will rarely change.