Add hack to work around issue with shadow DOM and @property
In preparation for the migration to Tailwind v4, add a workaround for
https://github.com/tailwindlabs/tailwindcss/issues/15005.

@property declarations are ignored by current browsers in Shadow DOM. Tailwind
v4's styles rely on these properties. In the Hypothesis client this noticeably
affects borders of elements in the sidebar container and shadows on the adder.

This commit works around this by duplicating @property declarations from
annotator.css inside the main document. In order to make this possible, the
annotator.css stylesheet must be loaded with the `crossOrigin` property set so
that it can be read by JS code.