Recalculate sidebar position when annotator styles load
If the annotator.css bundle loads after the sidebar is opened, the `margin-left`
property on the `.sidebar-container` element can be set incorrectly. Fix this
problem by invoking `Sidebar._onResize` when styles finish loading. This is the
same way we handle "resize" events on the window.

Part of https://github.com/hypothesis/client/issues/7307.