Fix sidebar font size in Chrome extension
The migration to Tailwind v4 resulted in the CSS rules defining the body's
font-family and font-size being moved to the base layer. This gave the rules
lower priority than some default styles which Chrome injects into all HTML pages
served from an extension (see links in code). As a result in the extension the
font size was being set to 12px (or 75% of the default) rather than our default
of 13px.

Work around this by changing these specific styles to be unlayered, like they
were when using Tailwind v3.