Protect crossorigin attribute of annotator.css stylesheet
pywb's frontend JS has logic that removes "crossorigin" attributes from `<link>`
elements [^1]. This causes problems for annotator.css because this stylesheet is
fetched cross-origin (we exclude it from being proxied) and we need to read the
response in JS to apply a workaround for `@property` and shadow DOM. To prevent
this, override `Element.prototype.removeAttribute` with a no-op function.

This fixes an issue where the annotator.css was loaded non-cross-origin in Via,
making the response opaque and preventing our `@property` workaround from being
applied, breaking some styles in the adder and annotator toolbar.

[^1]: https://github.com/webrecorder/wombat/blob/7433dede629b1c919c4c9c1e2c2daf1ac6665973/src/wombat.js#L2422